fixes mypy and split error module
