+++ services/web/server/tests/unit/with_dbs/03/invitations/test_invitations.py
@@ -16,11 +16,11 @@ from simcore_service_webserver.invitations._client import (
     InvitationsServiceApi,
     get_invitations_service_api,
 )
-from simcore_service_webserver.invitations.plugin import (
+from simcore_service_webserver.invitations.errors import (
     InvalidInvitation,
     InvitationsServiceUnavailable,
-    validate_invitation_url,
 )
+from simcore_service_webserver.invitations.plugin import validate_invitation_url
 from yarl import URL
